wip feat: create passable cross-chain message
