+++ apps/bridge/src/pages/Bridge/components/TransactionModal/ConfirmationStep.tsx
@@ -1,16 +1,17 @@
 import { Flex, FlexProps, LoadingSpinner, P } from '@interlay/ui';
+import { CrossChainTransferMessage } from '../../../../types/cross-chain';
 
 import { BridgeDetails } from '../BridgeDetails';
 import { TransactionDetails } from '../TransactionDetails';
 
 // eslint-disable-next-line @typescript-eslint/ban-types
-type Props = {};
+type Props = {message: CrossChainTransferMessage};
 
 type InheritAttrs = Omit<FlexProps, keyof Props | 'children'>;
 
 type ConfirmationStepProps = Props & InheritAttrs;
 
-const ConfirmationStep = ({ ...props }: ConfirmationStepProps): JSX.Element | null => {
+const ConfirmationStep = ({ message, ...props }: ConfirmationStepProps): JSX.Element | null => {
   return (
     <Flex alignItems='center' direction='column' gap='spacing8' {...props}>
       <LoadingSpinner color='secondary' diameter={100} thickness={8} variant='indeterminate' />
@@ -19,7 +20,7 @@ const ConfirmationStep = ({ ...props }: ConfirmationStepProps): JSX.Element | nu
         <P align='center' size='xs' weight='medium'>
           Please Confirm transfer in wallet.
         </P>
-        <TransactionDetails />
+        <TransactionDetails message={message} />
       </Flex>
     </Flex>
   );
