+++ apps/bridge/src/pages/Bridge/components/TransactionModal/TransactionModal.tsx
@@ -1,36 +1,24 @@
-import { useState } from 'react';
 import { CTA, Modal, ModalBody, ModalFooter, ModalHeader, ModalProps } from '@interlay/ui';
 
 import { ConfirmationStep } from './ConfirmationStep';
 import { SubmittedStep } from './SubmittedStep';
+import { CrossChainTransferMessage } from '../../../../types/cross-chain';
 
-enum Steps {
-  CONFIRMATION,
-  SUBMITTED
-}
-
-// eslint-disable-next-line @typescript-eslint/ban-types
-type Props = {};
+type Props = { message: CrossChainTransferMessage };
 
 type InheritAttrs = Omit<ModalProps, keyof Props | 'children'>;
 
 type TransactionModalProps = Props & InheritAttrs;
 
-const TransactionModal = ({ onClose, ...props }: TransactionModalProps): JSX.Element | null => {
-  // TODO: apply step progression
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [step, setStep] = useState(Steps.CONFIRMATION);
-
+const TransactionModal = ({ message, onClose, ...props }: TransactionModalProps): JSX.Element | null => {
   return (
     <Modal onClose={onClose} {...props}>
-      <ModalHeader align='start'>
-        {step === Steps.CONFIRMATION ? 'Waiting for confirmation' : 'Transaction Submitted'}
-      </ModalHeader>
+      <ModalHeader align='start'>{!message.status ? 'Waiting for confirmation' : 'Transaction Submitted'}</ModalHeader>
       <ModalBody>
-        {step === Steps.CONFIRMATION && <ConfirmationStep />}
-        {step === Steps.SUBMITTED && <SubmittedStep />}
+        {!message.status && <ConfirmationStep message={message} />}
+        {message.status && <SubmittedStep />}
       </ModalBody>
-      {step === Steps.SUBMITTED && (
+      {message.status && (
         <ModalFooter>
           <CTA onPress={onClose}>Close</CTA>
         </ModalFooter>
