+++ apps/bridge/src/pages/Bridge/hooks/useEthersProvider.ts
@@ -1,0 +1,25 @@
+import { useMemo } from 'react';
+import { type PublicClient, usePublicClient } from '@gobob/wagmi';
+import { FallbackProvider, JsonRpcProvider } from '@ethersproject/providers';
+
+import { type HttpTransport } from 'viem';
+
+export function publicClientToProvider(publicClient: PublicClient) {
+  const { chain, transport } = publicClient;
+  const network = {
+    chainId: chain.id,
+    name: chain.name,
+    ensAddress: chain.contracts?.ensRegistry?.address
+  };
+  if (transport.type === 'fallback')
+    return new FallbackProvider(
+      (transport.transports as ReturnType<HttpTransport>[]).map(({ value }) => new JsonRpcProvider(value?.url, network))
+    );
+  return new JsonRpcProvider(transport.url, network);
+}
+
+/** Hook to convert a viem Public Client to an ethers.js Provider. */
+export function useEthersProvider({ chainId }: { chainId?: number } = {}) {
+  const publicClient = usePublicClient({ chainId });
+  return useMemo(() => publicClientToProvider(publicClient), [publicClient]);
+}
