+++ apps/bridge/src/pages/Bridge/hooks/useEthersSigner.ts
@@ -1,0 +1,22 @@
+import { useMemo } from 'react';
+
+import { useWalletClient, type WalletClient } from '@gobob/wagmi';
+import { Web3Provider } from '@ethersproject/providers';
+
+export function walletClientToSigner(walletClient: WalletClient) {
+  const { account, chain, transport } = walletClient;
+  const network = {
+    chainId: chain.id,
+    name: chain.name,
+    ensAddress: chain.contracts?.ensRegistry?.address
+  };
+  const provider = new Web3Provider(transport, network);
+  const signer = provider.getSigner(account.address);
+  return signer;
+}
+
+/** Hook to convert a viem Wallet Client to an ethers.js Signer. */
+export function useEthersSigner({ chainId }: { chainId?: number } = {}) {
+  const { data: walletClient } = useWalletClient({ chainId });
+  return useMemo(() => (walletClient ? walletClientToSigner(walletClient) : undefined), [walletClient]);
+}
