+++ apps/bridge/src/types/common.ts
@@ -1,0 +1,3 @@
+type HexString = `0x${string}`;
+
+export type { HexString };
